refactor: statesをcontextフォルダに変更＆３つのuseContextのルート変更中_#25
